### 2026-09-03 37aecbc
Add OSV-Scanner workflow for vulnerability scanning
This workflow sets up periodic OSV-Scanner scanning for vulnerabilities and checks PRs for new vulnerabilities.
