+++ articles/textgrid-workflows.md
@@ -324,7 +324,7 @@ tg <- TextGrid$new(tg_file)
 load_time <- as.numeric(difftime(Sys.time(), start_time, units = "secs"))
 
 cat(sprintf("Loaded in %.3f seconds\n", load_time))
-#> Loaded in 0.009 seconds
+#> Loaded in 0.010 seconds
 cat(sprintf("Duration: %.2f minutes\n", tg$get_total_duration() / 60))
 #> Duration: 1.00 minutes
 cat(sprintf("File size: %.1f MB\n", file.size(tg_file) / 1024^2))
