+++ reference/PitchModule.md
@@ -87,13 +87,13 @@ pitch <- sound$to_pitch()
 pitch$duration
 #> function (...) 
 #> method(x, ...)
-#> <bytecode: 0x55f46307d288>
-#> <environment: 0x55f46094c730>
+#> <bytecode: 0x55c6bbb56538>
+#> <environment: 0x55c6bac2c150>
 pitch$nx
 #> function (...) 
 #> method(x, ...)
-#> <bytecode: 0x55f46307d288>
-#> <environment: 0x55f4602f1aa0>
+#> <bytecode: 0x55c6bbb56538>
+#> <environment: 0x55c6b9003888>
 
 # Query methods
 pitch$get_mean(0, 0, "hertz")  # Mean F0 in Hz
