+++ pkgdown.yml
@@ -22,7 +22,7 @@ articles:
   articles/v4-migration-guide: v4-migration-guide.html
   visualization: visualization.html
   vowel-space-analysis: vowel-space-analysis.html
-last_built: 2026-09-03T04:47Z
+last_built: 2026-09-03T10:50Z
 urls:
   reference: https://humlab-speech.github.io/pladdrr/reference
   article: https://humlab-speech.github.io/pladdrr/articles
