### 2026-09-03 49f37d0
Add pkgaudit workflow
Co-authored-by: FredrikKarlssonSpeech <200246+FredrikKarlssonSpeech@users.noreply.github.com>
