### 2026-09-04 #806 @adamhsparks
Hi @FredrikKarlssonSpeech, thank you for your submission. I see that there are some issues with the bot check. Can you check these and request the bot rerun the checks, `@ropensci-bot check package` will do the trick once you've fixed the package up.
