### 2026-09-04 #806 @FredrikKarlssonSpeech
This issue seems to be related to the bot not pulling in submodules. Would this prevent a review of the package?
