### 2026-09-04 4400f5d
Vendor pocketfft, drop unused pffft submodule; bump agent doc versions
rOpenSci's review bot clones without git submodule init, so
src/pffft and src/pocketfft (unlike src/praat.github.io, which the
package's own CI initializes via submodules: recursive) were empty
on their infra — though the actual reported build failure was in
src/praat.github.io/kar, left as a submodule per instruction pending
further assessment.

- pffft: no #include, no Makevars reference anywhere in the tree
  (.Rbuildignore already flagged it dead) — removed entirely.
- pocketfft: is the FFT backend for NUMFourier.cpp — 
[…]
