+++ reference/LongSound.md
@@ -109,7 +109,7 @@ Sound$create_tone(frequency = 150, duration = 1.0)$save(tmp)
 ls <- LongSound$open(tmp)
 print(ls)
 #> <Praat LongSound>
-#>   File: /tmp/RtmpJQ7vzw/file23a5436c7897.wav 
+#>   File: /tmp/Rtmpk0hTjh/file23573d16d6c7.wav 
 #>   Duration: 1.000 seconds
 #>   Sample rate: 44100 Hz
 #>   Channels: 1 
