### 2026-09-05 #806 @adamhsparks
@mpadge, can you have a look at this and comment please?
