### 2026-09-05 #806 @adamhsparks
Hi, @FredrikKarlssonSpeech.

I've checked with @mpadge. Your assumption is correct about the build process.

However, there are issues with the current structure that should still be addressed that this has flagged.

1. *src/pffft is unused* — no reference to it anywhere outside its own submodule
   directory (no #include, no mention in Makevars*). It looks like a leftover from
   before the project switched to pocketfft (commit c67d7495, "replace Praat FFTPACK
   with pocketfft") and was never removed. This one should just be deleted.

2. *src/pocketfft is used* (-Ipocketfft in [Makevars.in](http://makevars.in/)/Makevars.win), but it's a
   *header-only* library per that same commit's message. Vendoring a single header file
   as a full git submodule is overkill — it could just be copied into the repo directly.

3. *src/praat.github.io* is the real justification: it's the actual Praat C++ source
   that the wrapper code compiles against. This is the one case where a submodule (pinned
   to a stable commit) makes some sense, since upstream Praat is large and updated
   independently.

4. The original speckit design doc (specs/001-praat-r-access/research.md) actually
   recommended the opposite of the current approach — vendored/copied Praat source,
   explicitly flagged as "more CRAN-friendly" — precisely because CRAN's R CMD build/
   check won't fetch submodule contents. A fresh git clone without
   --recurse-submodules (or git submodule update --init) currently leaves all three
   submodule directories empty, which breaks the build for anyone who doesn't know to
   initialize them. If CRAN submission is ever a goal, all three submodules would need to
   become vendored source rather than submodules.

Can you please address these points?
