### 2026-09-05 cb01ac0
Vendor src/praat.github.io: convert last submodule to tracked files
rOpenSci's review bot clones without submodule init, so the vendored
Praat sources never showed up and the build failed
(no rule to make target 'praat.github.io/kar/longchar.o'). pffft/pocketfft
were already converted; this does the same for praat.github.io, the last
remaining submodule, removing the structural dependency on submodule
init entirely.

- Deinit + git rm --cached the submodule, re-add its working tree
  (~3800 files) as ordinary tracked content pinned at the previous
  submodule commit (f36024531).
- Remove .gitm
[…]
