+++ src/praat.github.io/dwtools/Sound_and_Spectrogram_extensions.cpp
@@ -1,0 +1,424 @@
+/* Sound_and_Spectrogram_extensions.cpp
+ *
+ * Copyright (C) 1993-2024 David Weenink
+ *
+ * This code is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or (at
+ * your option) any later version.
+ *
+ * This code is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this work. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+/*
+ djmw 20010718
+ djmw 20020813 GPL header.
+ djmw 20041124 Changed call to Sound_to_Spectrum.
+ djmw 20070103 Sound interface changes
+ djmw 20071107 Errors/warnings text changes
+ djmw 20071202 Melder_warning<n>
+*/
+
+#include "Sound_and_Spectrogram_extensions.h"
+#include "Sound_extensions.h"
+#include "Sound_and_Spectrum.h"
+#include "Sound_to_Pitch.h"
+#include "Vector.h"
+#include "NUM2.h"
+
+// Forward declarations for SIMD MFCC operations
+#ifdef HAVE_XSIMD
+extern "C" {
+	double triangular_filter_simd_bridge(
+		constVEC const& spectrum_power,
+		constVEC const& frequencies,
+		integer ifrom,
+		integer ito,
+		double fl_hz,
+		double fc_hz,
+		double fh_hz
+	);
+	bool should_use_simd_for_mfcc();
+}
+#endif
+
+autoSound BandFilterSpectrogram_as_Sound (BandFilterSpectrogram me, int to_dB);
+
+/*
+	The gaussian(x) = (exp(-48*((i-(n+1)/2)/(n+1))^2)-exp(-12))/(1-exp(-12));
+	For power we need the area under the square of this window:
+	Integrate (gaussian(i)^2,i=1..n) =
+		(sqrt(Pi)*sqrt(3)*sqrt(2)*erf(2*(n-1)*sqrt(3)*sqrt(2)/(n+1))*(n+1) + 24*exp(-24)*(n-1)+
+		-4*sqrt(Pi)*sqrt(3)*exp(-12)*erf(2*(n-1)*sqrt(3)/(n+1))*(n+1))/ (24 * (-1+exp(-12))^2),
+	where erf(x) = 1 - erfc(x) and n is the windowLength in samples.
+	To compare with the rectangular window we need to divide this by the window width (n -1) x 1^2.
+*/
+static void _Spectrogram_windowCorrection (Spectrogram me, integer numberOfSamples_window) {
+	double windowFactor = 1.0;
+	if (numberOfSamples_window > 1) {
+		const double e12 = exp (-12);
+		const double denum = (e12 - 1) * (e12 - 1.0) * 24 * (numberOfSamples_window - 1);
+		const double arg1 = 2.0 * NUMsqrt3 * (numberOfSamples_window - 1) / (numberOfSamples_window + 1);
+		const double arg2 = arg1 * NUMsqrt2;
+		const double p2 = NUMsqrtpi * NUMsqrt3 * NUMsqrt2 * (1 - NUMerfcc (arg2)) * (numberOfSamples_window + 1);
[…]
