+++ src/praat.github.io/fon/praat_Tiers.cpp
@@ -1,0 +1,2255 @@
+/* praat_Tiers.cpp
+ *
+ * Copyright (C) 1992-2018,2020-2025 Paul Boersma
+ *
+ * This code is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 3 of the License, or (at
+ * your option) any later version.
+ *
+ * This code is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this work. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "H1minusH2Tier.h"
+#include "Ltas.h"
+#include "PitchTier_to_PointProcess.h"
+#include "PitchTier_to_Sound.h"
+#include "Matrix_and_PointProcess.h"
+#include "PointProcess_and_Sound.h"
+#include "Sound_PointProcess.h"
+#include "SpectrumTier.h"
+#include "TextGrid.h"
+#include "VoiceAnalysis.h"
+
+#include "praat_Tiers.h"
+
+// MARK: - AMPLITUDETIER
+
+// MARK: New
+
+FORM (NEW1_AmplitudeTier_create, U"Create empty AmplitudeTier", nullptr) {
+	WORD (name, U"Name", U"empty")
+	REAL (startTime, U"Start time (s)", U"0.0")
+	REAL (endTime, U"End time (s)", U"1.0")
+	OK
+DO
+	Melder_require (endTime > startTime,
+		U"The end time should be greater than the start time.");
+	CREATE_ONE
+		autoAmplitudeTier result = AmplitudeTier_create (startTime, endTime);
+	CREATE_ONE_END (name)
+}
+
+// MARK: Help
+
+DIRECT (HELP_AmplitudeTier_help) {
+	HELP (U"AmplitudeTier")
+}
+
+// MARK: View & Edit
+
+DIRECT (EDITOR_ONE_WITH_ONE_AmplitudeTier_viewAndEdit) {
+	EDITOR_ONE_WITH_ONE (an,AmplitudeTier, Sound)   // Sound may be null
+		autoAmplitudeTierEditor editor = AmplitudeTierEditor_create (ID_AND_FULL_NAME, me, you);
+	EDITOR_ONE_WITH_ONE_END
+}
+
+DIRECT (HINT_AmplitudeTier_Sound_viewAndEdit) {
+	INFO_NONE
+		Melder_information (U"To include a copy of a Sound in your AmplitudeTier window:\n"
+			"   select an AmplitudeTier and a Sound, and click \"View & Edit\".");
+	INFO_NONE_END
+}
+
+// MARK: Query
+
+FORM (REAL_AmplitudeTier_getShimmer_local, U"AmplitudeTier: Get shimmer (local)", U"AmplitudeTier: Get shimmer (local)...") {
+	REAL (shortestPeriod, U"Shortest period (s)", U"0.0001")
+	REAL (longestPeriod, U"Longest period (s)", U"0.02")
+	POSITIVE (maximumAmplitudeFactor, U"Maximum amplitude factor", U"1.6")
+	OK
+DO
+	QUERY_ONE_FOR_REAL (AmplitudeTier)
[…]
