+++ reference/PitchModule.md
@@ -87,13 +87,13 @@ pitch <- sound$to_pitch()
 pitch$duration
 #> function (...) 
 #> method(x, ...)
-#> <bytecode: 0x55a7f81006a0>
-#> <environment: 0x55a7f89ebe80>
+#> <bytecode: 0x55b551cecb20>
+#> <environment: 0x55b553c77720>
 pitch$nx
 #> function (...) 
 #> method(x, ...)
-#> <bytecode: 0x55a7f81006a0>
-#> <environment: 0x55a7f91fcd18>
+#> <bytecode: 0x55b551cecb20>
+#> <environment: 0x55b5537ab198>
 
 # Query methods
 pitch$get_mean(0, 0, "hertz")  # Mean F0 in Hz
