+++ llms.txt
@@ -501,9 +501,9 @@ Please also cite Praat itself:
 
 Contributions are welcome. See
 [CONTRIBUTING.md](https://github.com/humlab-speech/pladdrr/blob/main/.github/CONTRIBUTING.md)
-for how to get a working build (the vendored Praat sources are git
-submodules, and GSL is required) and what a good pull request looks
-like.
+for how to get a working build (the vendored Praat sources are tracked
+directly in this repo, and GSL is required) and what a good pull request
+looks like.
 
 Please note that this project is released with a [Contributor Code of
 Conduct](https://github.com/humlab-speech/pladdrr/blob/main/.github/CODE_OF_CONDUCT.md).
