+++ articles/formantpath-robust-tracking.md
@@ -505,7 +505,7 @@ system.time(frm_result <- sound$to_formant_burg())
 #>   0.018   0.000   0.009
 system.time(fp <- sound$to_formant_path(num_steps_up_down = 2L))
 #>    user  system elapsed 
-#>   0.069   0.002   0.042
+#>   0.068   0.001   0.039
 ```
 
 ## Best Practices
