### 2026-09-05 #806 @FredrikKarlssonSpeech
Thanks for the feedback.

I see your reasoning and agree. I have cleaned up the submodules, and they are now just ordinary tracked files.

On the duplicate function names, those are intentional. They match Praat's own object-class names (Formant, LPC, Table, etc.) so that R code reads like the equivalent Praat script. The purpose of the pladdrr package is to make it easy to re-implement Praat code in efficient R, so it is preferable to keep Praat's object naming.

This is a relatively large package, and we have not reached 100% test coverage. It currently sits at 82% according to codecov.

The version numbers in AGENT_GUIDE.md/PRAAT_MODIFICATIONS.md have been bumped to match DESCRIPTION.

The comments above reflect the most recently pushed commit for this package
https://github.com/humlab-speech/pladdrr/commit/cb01ac0208ea906cc9842538b2c94b08102be47c

and I ask you to run the automatic evaluation again.

@ropensci-bot check package
