### 2026-09-05 63a1355
Add Semgrep workflow for code scanning
This workflow integrates Semgrep for code scanning on pushes and pull requests to the main branch, with a scheduled job for regular scans.
