### 2026-09-05 f098c6b
Add OSSAR workflow for static analysis
This workflow integrates OSSAR for static analysis and uploads results to the Security tab.
