+++ articles/performance-optimization.md
@@ -312,7 +312,7 @@ system.time({
   stats <- get_pitch_stats_direct(pitch_ptr)
 })
 #>    user  system elapsed 
-#>   0.011   0.000   0.004
+#>   0.010   0.000   0.004
 # fewer crossings; measure on your own data
 ```
 
@@ -331,7 +331,7 @@ system.time({
   pitches <- lapply(sounds, function(s) s$to_pitch())
 })
 #>    user  system elapsed 
-#>   0.206   0.009   0.101
+#>   0.206   0.009   0.102
 # measure on your own data
 
 # Tier 3: Batch
@@ -340,7 +340,7 @@ system.time({
   pitches <- sound_to_pitch_batch(sounds)
 })
 #>    user  system elapsed 
-#>   0.202   0.008   0.096
+#>   0.209   0.002   0.096
 # scales with cores; measure on your own data
 
 # Tier 3: Parallel (2 cores)
@@ -349,7 +349,7 @@ system.time({
 })
 #> Processing 20 files using 2 cores (2 thread(s)/worker)
 #>    user  system elapsed 
-#>   0.099   0.088   0.373
+#>   0.101   0.080   0.371
 # includes file I/O; measure on your own data
 ```
 
@@ -368,7 +368,7 @@ system.time({
   }, numeric(1))
 })
 #>    user  system elapsed 
-#>   0.012   0.000   0.012
+#>   0.013   0.000   0.012
 # one R->C crossing per time point
 
 # Tier 3: Vectorized
@@ -466,9 +466,9 @@ benchmark_results <- benchmark_parallel(
 #> Processing 10 files using 2 cores (2 thread(s)/worker)
 
 print(benchmark_results)
-#>   cores   time_sec   speedup
-#> 1     1 0.05282569 1.0000000
-#> 2     2 0.06043959 0.8740247
+#>   cores   time_sec  speedup
+#> 1     1 0.05262661 1.000000
+#> 2     2 0.06242895 0.842984
 # Inspect the returned table to see where returns diminish on your machine
 ```
 
