+++ reference/longsound_open.md
@@ -28,7 +28,7 @@ Sound$create_tone(frequency = 220, duration = 1, sampling_rate =
 ls <- longsound_open(wav)
 print(ls)
 #> <Praat LongSound>
-#>   File: /tmp/RtmpDHY0T5/file227ed09376b.wav 
+#>   File: /tmp/RtmpOcvedA/file22ef8f6176d.wav 
 #>   Duration: 1.000 seconds
 #>   Sample rate: 16000 Hz
 #>   Channels: 1 
