+++ pkgdown.yml
@@ -22,7 +22,7 @@ articles:
   articles/v4-migration-guide: v4-migration-guide.html
   visualization: visualization.html
   vowel-space-analysis: vowel-space-analysis.html
-last_built: 2026-09-05T05:51Z
+last_built: 2026-09-05T06:48Z
 urls:
   reference: https://humlab-speech.github.io/pladdrr/reference
   article: https://humlab-speech.github.io/pladdrr/articles
