+++ reference/PitchModule.md
@@ -87,13 +87,13 @@ pitch <- sound$to_pitch()
 pitch$duration
 #> function (...) 
 #> method(x, ...)
-#> <bytecode: 0x55d235ce2b40>
-#> <environment: 0x55d235c4ab10>
+#> <bytecode: 0x563d8b1e5520>
+#> <environment: 0x563d8a8409a0>
 pitch$nx
 #> function (...) 
 #> method(x, ...)
-#> <bytecode: 0x55d235ce2b40>
-#> <environment: 0x55d234b04e78>
+#> <bytecode: 0x563d8b1e5520>
+#> <environment: 0x563d8d60a268>
 
 # Query methods
 pitch$get_mean(0, 0, "hertz")  # Mean F0 in Hz
