+++ reference/LongSound.md
@@ -109,7 +109,7 @@ Sound$create_tone(frequency = 150, duration = 1.0)$save(tmp)
 ls <- LongSound$open(tmp)
 print(ls)
 #> <Praat LongSound>
-#>   File: /tmp/RtmpOcvedA/file22ef89cd2b4.wav 
+#>   File: /tmp/RtmpRc9hLe/file229e29bb4bad.wav 
 #>   Duration: 1.000 seconds
 #>   Sample rate: 44100 Hz
 #>   Channels: 1 
