+++ reference/benchmark_parallel.md
@@ -50,7 +50,7 @@ results <- benchmark_parallel(
 #> Processing 1 files using 2 cores (2 thread(s)/worker)
 print(results)
 #>   cores    time_sec   speedup
-#> 1     1 0.002428532 1.0000000
-#> 2     2 0.005863905 0.4141492
+#> 1     1 0.002213717 1.0000000
+#> 2     2 0.005964041 0.3711773
 # }
 ```
