### 2026-09-05 #806 @adamhsparks
@ropensci-reviewbot check package
