### 2026-09-05 #806 @adamhsparks
I see this note, “ NOTE: Some imported packages appear to have no associated function calls; please ensure with author that these 'Imports' are listed appropriately”.

Have you run `R CMD check` and can confirm that this is in error?

Will start looking for a handling editor now.
