### 2026-09-06 #806 @FredrikKarlssonSpeech
> I see this note, “ NOTE: Some imported packages appear to have no associated function calls; please ensure with author that these 'Imports' are listed appropriately”.
>
> Have you run `R CMD check` and can confirm that this is in error?

Yes this note is likelly due to the rlang .data import ("#' @importFrom rlang .data" in pladdrr-package.R ) used to prepare the use of .data$ in the plotting functions. So, rlang is used, while I am sure it can appear to not be.
